requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==41.0.2
debugpy==1.6.7
dj-rest-auth==4.0.1
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
gunicorn==21.2.0
importlib-metadata==6.8.0
importlib-resources==6.0.0
inflection==0.5.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
Markdown==3.4.4
mysqlclient==2.2.0
packaging==23.1
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
rpds-py==0.9.2
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing-extensions==4.7.1
uritemplate==4.1.1
zipp==3.16.2
